added row parser to the geoHelper
