+++ src/js/components/search/visualizations/geo/GeoVisualizationSection.jsx
@@ -290,8 +290,7 @@ export default class GeoVisualizationSection extends React.Component {
                     center={[-95.569430, 38.852892]}
                     className={this.props.className}
                     mapLegendToggle={this.props.mapLegendToggle}
-                    updateMapLegendToggle={this.props.updateMapLegendToggle}
-                    prohibitedCountryCodes={this.props.prohibitedCountryCodes} >
+                    updateMapLegendToggle={this.props.updateMapLegendToggle} >
                     {message}
                 </MapWrapper>
                 <Note message={noteMessage} />
