+++ src/js/containers/search/visualizations/geo/GeoVisualizationSectionContainer.jsx
@@ -22,8 +22,7 @@ import MapBroadcaster from 'helpers/mapBroadcaster';
 import Analytics from 'helpers/analytics/Analytics';
 
 import SearchAwardsOperation from 'models/v1/search/SearchAwardsOperation';
-import BaseSpendingByGeographyResult
-    from "../../../../models/v2/search/visualizations/geo/BaseSpendingByGeographyResult";
+import { parseRows } from 'helpers/search/visualizations/geoHelper';
 
 const propTypes = {
     reduxFilters: PropTypes.object,
@@ -61,8 +60,6 @@ const logMapScopeEvent = (scope) => {
     });
 };
 
-// FABS prohibits the following US territories or freely associated states from being submitted using their GENC code.  They should be submitted using the country code "USA"
-const prohibitedCountryCodes = ['ASM', 'FSM', 'GUM', 'MHL', 'MNP', 'PLW', 'PRI', 'VIR', 'XBK', 'XHO', 'XJA', 'XJV', 'XKR', 'XMW', 'XNV', 'XPL', 'XWK'];
 
 export class GeoVisualizationSectionContainer extends React.Component {
     constructor(props) {
@@ -253,41 +250,14 @@ export class GeoVisualizationSectionContainer extends React.Component {
             error: false
         });
 
-        const addUSTerritories = (results) => {
-            const filteredArray = apiParams.geo_layer_filters.filter((value) => prohibitedCountryCodes.includes(value));
-            const usaData = results.find((value) => value.shape_code === 'USA');
-
-            if (usaData) {
-                filteredArray.forEach((value) => {
-                    results.push({
-                        shape_code: value,
-                        display_name: usaData.display_name,
-                        aggregated_amount: usaData.aggregated_amount,
-                        per_capita: usaData.per_capita,
-                        population: usaData.population
-                    });
-                });
-            }
-
-            return results;
-        };
-
 
         this.props.setAppliedFilterCompletion(false);
         this.apiRequest = SearchHelper.performSpendingByGeographySearch(apiParams);
         this.apiRequest.promise
             .then((res) => {
                 this.apiRequest = null;
-                const data = res.data.results;
-                const parsedData = data.map((item) => {
-                    const geoElement = Object.create(BaseSpendingByGeographyResult);
-                    geoElement.populate(item);
-
-                    return geoElement;
-                });
-
-                const allGENCCodes = apiParams.geo_layer === "country" ? addUSTerritories(parsedData) : parsedData;
-                this.setState({ rawAPIData: allGENCCodes }, this.setParsedData);
+                const parsedData = parseRows(res.data.results, apiParams);
+                this.setState({ rawAPIData: parsedData }, this.setParsedData);
             })
             .catch((err) => {
                 if (!isCancel(err)) {
@@ -373,8 +343,7 @@ export class GeoVisualizationSectionContainer extends React.Component {
                 mapLegendToggle={this.props.mapLegendToggle}
                 subaward={this.props.subaward}
                 isDefCodeInFilter={this.props.reduxFilters?.defCodes?.counts}
-                className={this.props.className}
-                prohibitedCountryCodes={prohibitedCountryCodes} />
+                className={this.props.className} />
         );
     }
 }
