+++ src/js/models/v2/search/visualizations/geo/BaseSpendingByGeographyResult.js
@@ -3,7 +3,7 @@
  * Created by Andrea Blackwell 12/14/2023
  */
 
-import * as MoneyFormatter from 'helpers/moneyFormatter';
+import { formatMoneyWithPrecision } from 'helpers/moneyFormatter';
 
 const BaseSpendingByGeographyResult = {
     populate(data) {
@@ -12,6 +12,12 @@ const BaseSpendingByGeographyResult = {
         this.aggregated_amount = data.aggregated_amount || 0;
         this.population = data.population || 0;
         this.per_capita = data.per_capita || 0;
+    },
+    get aggregatedAmountFormatted() {
+        return formatMoneyWithPrecision(this.aggregated_amount, 0);
+    },
+    get perCapitaAggregated() {
+        return formatMoneyWithPrecision(this.per_capita, 0);
     }
 };
